refactor: split the TCC diagnostics and the failure log into leaf modules
server.py had grown to 4,537 lines. Two of its sections are genuine leaves —
they call nothing else in the file — so they move out into their own modules,
and nothing else does:

- tcc.py — the macOS protected-folder detection and its guidance message.
- failure_log.py — the opt-in COMFY_LOCAL_MCP_DEBUG_LOG log: its config, its
  module state (_FAILURE_LOG_PATH, the lazily-opened rotating handler) and
  _log_failure.
- textutil.py — the pure text helpers both server and failure_log need
  (_tail / _stream_tail / _redact_url). Extracted so failure_log can be a real
  leaf instead of importing back into server.

The three moved bodies are byte-identical to their former selves; the only
edits inside server.py are the import line and requalified call sites.

server reaches the new modules module-qualified (tcc._tcc_guidance(...),
failure_log._log_failure(...)) and re-exports nothing. That is the point: a
test that kept patching a moved name on server would otherwise silently patch a
name nothing reads, and pass while testing nothing. With no re-exports the
stale patch raises AttributeError instead, which is how every patch site in
conftest.py and test_failure_log.py was found and repointed at its owning
module.

The envelope core, the --json-stream machinery, spend consent, the version /
compat probes and all ~24 tools stay in server.py.
